+++ src/main/java/lotto/view/InputView.java
@@ -3,7 +3,7 @@ package lotto.view;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
-import lotto.domain.utils.Parser;
+import lotto.utils.Parser;
 
 public class InputView {
